Take, back, drop commands & other improvements.
Changes:

1) Spelling & wording improvements across entire project + javadoc
stuff. Let me know if I missed anything or feel free to change it
yourself as you notice it.

2) Updated player class to handle previous room & inventory
appropriately.

3) added a printInfo() method that prints the room details as well as
the player's inventory, replaced getLongDescription() located in various
places within the Game class.

4) New commands: Take, back, & drop.
Take:
Take an item that is nearby in the room and add it to the player's
inventory if it exists.
Drop:
Drops an item into a room so long as the player has it.
Back: Goes back to previous room. Does not work on first turn nor when
the player is in jail (crude implementation for the jail check, could be
better). Also, changed the "go" command so that if a player types "Go
back" it will send the player back a room.