Added main function, matrix recursion works, added use and examine commands.
Added examine command:
Gives description and weight of item in a player's inventory.
Added use command:
Use any item in the player's inventory. Uses an if-elseif-else check for
whether or not an item has a special function.

Added main:
When starting the game, right click the Game class in blueJ and hit the
void main[] function and then press "OK" to start the game.

Matrix:
Matrix has been added! Not unlimited, when you visit the classroom in
the "matrix" version of the game, you go to the inTheMatrix room instead
of the computer room. Also added special dialog when in the matrix so
players do not get lost.