Added map and changed roomList
Adding certain rooms to the roomList array produced undesirable results:

For example, the bus can't drop you off in your own apartment or the
back of the grocery room. The matrix room HAS to be removed otherwise
you can get stuck in there if you currently aren't in the matrix. Lol.

Added the map in .png format.