Added usable beamer item + changed item locations
The beamer is now an item that can be found in the player's apartment -
it can be charged in a room and fired in a later room to be teleported
back to where it was charged. I also moved some items around, added
others, and removed the vault key in preparation for NPCs.